イベント詳細をモーダルで統一して表示するよう修正
