+++ src/App.jsx
@@ -2,19 +2,43 @@ import React, { useState, useEffect } from 'react'
 import CalendarView from './components/CalendarView'
 import SearchBar from './components/SearchBar'
 import SearchResults from './components/SearchResults'
+import EventModal from './components/EventModal'
 import { events } from './data/events'
 
 function App() {
   const [selectedDate, setSelectedDate] = useState(null)
+  const [selectedEvent, setSelectedEvent] = useState(null)
+  const [isModalOpen, setIsModalOpen] = useState(false)
   const [searchResults, setSearchResults] = useState([])
   const [query, setQuery] = useState('')
 
-  // 🔍 入力が空になったら検索結果をクリア
   useEffect(() => {
     if (query.trim() === '') {
       setSearchResults([])
     }
   }, [query])
+
+  const formatDate = (date) => {
+    const y = date.getFullYear()
+    const m = String(date.getMonth() + 1).padStart(2, '0')
+    const d = String(date.getDate()).padStart(2, '0')
+    return `${y}-${m}-${d}`
+  }
+
+  const handleDateSelect = (date) => {
+    setSelectedDate(date)
+
+    const formatted = formatDate(date)
+    const found = events.find(event => event.date === formatted)
+
+    if (found) {
+      setSelectedEvent(found)
+      setIsModalOpen(true)
+    } else {
+      setSelectedEvent(null)
+      setIsModalOpen(false)
+    }
+  }
 
   const handleSearch = (keyword) => {
     const trimmed = keyword.trim()
@@ -26,7 +50,6 @@ function App() {
     }
 
     const lowerKeyword = trimmed.toLowerCase()
-
     const filtered = events.filter(event =>
       event.title.toLowerCase().includes(lowerKeyword) ||
       event.description.toLowerCase().includes(lowerKeyword)
@@ -34,23 +57,6 @@ function App() {
 
     setSearchResults(filtered)
   }
-
-  // ✅ ローカル時間でフォーマットする関数（UTC問題対策）
-  const formatDate = (date) => {
-    const y = date.getFullYear()
-    const m = String(date.getMonth() + 1).padStart(2, '0')
-    const d = String(date.getDate()).padStart(2, '0')
-    return `${y}-${m}-${d}`
-  }
-
-  const handleDateSelect = (date) => {
-    setSelectedDate(date)
-  }
-
-  // 🗓️ 選択された日付に対応するイベントだけを抽出（ローカル時間で一致）
-  const selectedEvents = selectedDate
-    ? events.filter(event => event.date === formatDate(selectedDate))
-    : []
 
   return (
     <div className="main-container">
@@ -72,20 +78,22 @@ function App() {
 
       {selectedDate && (
         <div className="selected-date">
-          <strong>選択された日付：</strong> {selectedDate.toLocaleDateString('ja-JP', { weekday: 'short', year: 'numeric', month: 'short', day: 'numeric' })}
+          <strong>選択された日付：</strong>{' '}
+          {selectedDate.toLocaleDateString('ja-JP', {
+            weekday: 'short',
+            year: 'numeric',
+            month: 'short',
+            day: 'numeric'
+          })}
         </div>
       )}
 
-      {selectedEvents.length > 0 && (
-        <div className="event-details">
-          <h2>イベント詳細</h2>
-          {selectedEvents.map((event, index) => (
-            <div key={index} className="event-item">
-              <h3>{event.title}</h3>
-              <p>{event.description}</p>
-            </div>
-          ))}
-        </div>
+      {isModalOpen && selectedEvent && (
+        <EventModal
+          isOpen={isModalOpen}
+          onClose={() => setIsModalOpen(false)}
+          event={selectedEvent}
+        />
       )}
     </div>
   )
